[Refactor] Relocate to `domain.place.dto.response.member`
