+++ src/main/java/boombimapi/domain/place/dto/response/member/ResolveMemberPlaceResponse.java
@@ -1,4 +1,4 @@
-package boombimapi.domain.place.dto.response;
+package boombimapi.domain.place.dto.response.member;
 
 import boombimapi.domain.place.entity.MemberPlace;
